Non active selectors no longer visible. Fixed errors on trying to reload the game without quitting

+++ Assets/Scripts/App/IOController.cs
@@ -14,7 +14,7 @@ namespace App
 
         public void DeleteDirectory(string relativePath)
         {
-            Assert.IsTrue(DoesRelativeDirectoryExist(relativePath));
+            if (!DoesRelativeDirectoryExist(relativePath)) return;
             Directory.Delete(Path.Combine(_appDataPath, relativePath), true);
         }
